Prevent playing card when player doesn't have enough mana.

+++ src/UI/handReducer.ts
@@ -6,13 +6,23 @@ import { Card, CardList } from '../Card';
 import { hand } from './initialState';
 import { summonMinion } from './boardReducer';
 import { fromCard } from '../Minion';
+import { spendMana } from './characterReducer';
+import { currentPlayer, Game } from '../Game';
+import { canSpendMana } from '../Hero';
 
 const actionCreator = actionCreatorFactory();
 export const removeCard = actionCreator<Card>('REMOVE_CARD');
-export const playCard = (
-  payload: Card
-): ThunkAction<void, {}, {}> => dispatch => {
+export const playCard = (payload: Card): ThunkAction<void, Game, {}> => (
+  dispatch,
+  getState
+) => {
+  const hero = currentPlayer(getState());
+  if (!canSpendMana(hero, payload.cost)) {
+    return;
+  }
+
   dispatch(removeCard(payload));
+  dispatch(spendMana(payload.cost));
 
   switch (payload.type) {
     case 'minion':
